⚡ Created new email based login and registering flow
🐞 Fixed google sign-in not working on iOS
